+++ src/components/sub-components/Object.tsx
@@ -24,7 +24,7 @@ const Object: React.FC<SubComponentProps> = ({ object, updateQuery, setObject, d
       try {
         let token = '';
         token = await datasource.getToken();
-        results = await datasource.sevOneConnection.getObjects(token, 3, device, deviceGroup, 20, 0, '');
+        results = await datasource.sevOneConnection.getObjects(token, 3, device, deviceGroup, 10000, 0, '');
       } catch (err) {
         console.error('Error Loading Devices: ', err);
         results = [{ label: 'Error Loading Devices', value: '' }];