+++ src/types.ts
@@ -22,7 +22,7 @@ export const DEFAULT_QUERY: Partial<MyQuery> = {
     description: 'Grab All Devices Info from SevOne',
   },
   page: 0,
-  size: 20,
+  size: 100,
   device: [],
   object: [],
   indicator: [],